-Wtype-defaults

+++ src/Data/IntMod.hs
@@ -82,7 +82,7 @@ newtype IntMod = IntMod {getIntMod :: Int}
   deriving newtype (Eq, Ord, Read, Show, Real, Prim)
 
 intMod :: (Integral a) => a -> IntMod
-intMod x = fromIntegral $ mod (fromIntegral x) MOD
+intMod x = fromIntegral $ mod x MOD
 {-# INLINE intMod #-}
 
 intModValidate :: IntMod -> Bool
